Allow specifying custom host + port
In my own usage of this action, I have found a specific case where I need to run a step inside a container, but the turborepo cache server on a host machine. In this case, without being able to specify a custom host+port, there's no way for turbo tasks running inside the container to be able to communicate with the cache server. While these inputs are not normally needed, they subtly enhance the flexibility of this action.